refactor(cartForm component): making a stateless tmpl for it
