+++ client/components/cartForm.js
@@ -2,28 +2,13 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { fetchCart, removeItem } from '../store/cart'
 import { me } from '../store/user'
-import Button from '@material-ui/core/Button'
-import Table from '@material-ui/core/Table'
-import TableBody from '@material-ui/core/TableBody'
-import TableCell from '@material-ui/core/TableCell'
-import TableHead from '@material-ui/core/TableHead'
-import TableRow from '@material-ui/core/TableRow'
-import TableFooter from '@material-ui/core/TableFooter'
-import Paper from '@material-ui/core/Paper'
-import Input from '@material-ui/core/Input'
-import { withStyles } from '@material-ui/core/styles'
 import Payment from './payment'
+import FormTmpl from './cartForm.tmpl'
 
-const CustomTableCell = withStyles(theme => ({
-  head: {
-    backgroundColor: 'dodgerblue',
-    color: theme.palette.common.white,
-  },
-  body: {
-    fontSize: 14,
-  },
-}))(TableCell)
-
+/**
+ * Class for users cart
+ * @extends Component
+ */
 class CartForm extends Component {
   constructor(props) {
     super(props)
@@ -50,9 +35,6 @@ class CartForm extends Component {
   }
 
   async handleChange(item, ev) {
-    // console.log("What we're typing: ", event.target.value)
-    // this.setState({ [event.target.name]: event.target.value });
-    //console.log(event.target.value, amount)
     let priceChangeCart = [...this.state.cart]
     let price = 0;
     for (let i = 0; i < priceChangeCart.length; i++) {
@@ -65,7 +47,6 @@ class CartForm extends Component {
         }
         price += +priceChangeCart[i].amount
     }
-    console.log(price)
     this.setState({
       amount: price
     })
@@ -80,63 +61,17 @@ class CartForm extends Component {
   }
 
   render() {
+      const propSetup = {
+          ...this.props,
+          handleSubmit: this.handleSubmit,
+          handleChange: this.handleChange,
+          handleDelet: this.handleDelete,
+          amount: this.state.amount,
+          description: this.state.description,
+          transactions: this.state.cart
+      }
     return (
-      <form onSubmit={this.handleSubmit}>
-        <Paper>
-          <Table>
-            <TableHead>
-              <TableRow>
-                <CustomTableCell>Opinion</CustomTableCell>
-                <CustomTableCell>Price</CustomTableCell>
-                <CustomTableCell />
-                <CustomTableCell />
-                <CustomTableCell />
-                <CustomTableCell />
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {this.props &&
-                this.props.cart &&
-                this.props.cart[0] &&
-                this.props.cart[0].id &&
-                this.props.cart.map((item) => (
-                  <TableRow key={item.id}>
-                    <TableCell>{item.opinion ? item.opinion.statement : 'Dat New New'}</TableCell>
-                    <TableCell>
-                    <TableCell><Input placeholder="What Is It Worth?" value={item.amount} onChange={ev => {this.handleChange(item, ev)}} /></TableCell>
-                    </TableCell>
-                    <TableCell>
-                      <Button
-                        variant="contained"
-                        color="secondary"
-                        type="submit"
-                        onClick={() => this.handleDelete(item.id)}
-                      >
-                        <small>Delete</small>
-                      </Button>
-                    </TableCell>
-                  </TableRow>
-                ))}
-            </TableBody>
-            <TableFooter>
-              <TableRow>
-                <TableCell />
-                <TableCell />
-                <TableCell />
-                <TableCell>Total</TableCell>
-                <TableCell>$ {this.state.amount}</TableCell>
-                <TableCell>
-                  <Payment
-                    amount={this.state.amount}
-                    description={this.state.description}
-                    transactions={this.state.cart}
-                  />
-                </TableCell>
-              </TableRow>
-            </TableFooter>
-          </Table>
-        </Paper>
-      </form>
+      <FormTmpl {...propSetup}/>
     )
   }
 }
